+++ src/chess.zig
@@ -4,6 +4,7 @@ const square = @import("chess/square.zig");
 pub const BitBoard = @import("chess/BitBoard.zig");
 pub const Color = color.Color;
 pub const File = square.File;
+pub const Piece = @import("chess/piece.zig").Piece;
 pub const Rank = square.Rank;
 pub const RankFile = square.RankFile;
 pub const Side = color.Side;