+++ src/tests.zig
@@ -6,5 +6,8 @@ comptime {
     _ = @import("chess/bit_board/LookupImpl.zig");
     _ = @import("chess/bit_board/MathImpl.zig");
     _ = @import("chess/color.zig");
+    _ = @import("chess/piece.zig");
+    _ = @import("chess/piece/ConditionImpl.zig");
+    _ = @import("chess/piece/LookupImpl.zig");
     _ = @import("chess/square.zig");
 }