+++ src/zaplum.zig
@@ -6,4 +6,5 @@ pub const options: Options = if (@hasDecl(root, "zaplum_options")) root.zaplum_o
 
 pub const Options = struct {
     bit_board_impl: chess.BitBoard.Impl = chess.BitBoard.default_impl,
+    piece_impl: chess.Piece.Impl = chess.Piece.default_impl,
 };